refactor: move static methods from interface to utils class
